Move AccountDoesNotExistError to domain
While `AccountDoesNotExistError` is an error that originates from the DataSource, these errors can and should also be used in the Domain layer.

Since the Domain layer is the one specifying the interface for the DataSource to implement, then specifying that the error can also be thrown by implementations of the interface, hints that it is a domain-specific error.
