+++ src/datasources/account/account.datasource.spec.ts
@@ -2,7 +2,7 @@ import { AccountDataSource } from '@/datasources/account/account.datasource';
 import * as postgres from 'postgres';
 import { PostgresError } from 'postgres';
 import { faker } from '@faker-js/faker';
-import { AccountDoesNotExistError } from '@/datasources/account/errors/account-does-not-exist.error';
+import { AccountDoesNotExistError } from '@/domain/account/errors/account-does-not-exist.error';
 import * as shift from 'postgres-shift';
 import configuration from '@/config/entities/__tests__/configuration';
 import { EmailAddress } from '@/domain/account/entities/account.entity';
