+++ src/datasources/account/account.datasource.ts
@@ -1,7 +1,7 @@
 import { Inject, Injectable } from '@nestjs/common';
 import * as postgres from 'postgres';
 import { IAccountDataSource } from '@/domain/interfaces/account.datasource.interface';
-import { AccountDoesNotExistError } from '@/datasources/account/errors/account-does-not-exist.error';
+import { AccountDoesNotExistError } from '@/domain/account/errors/account-does-not-exist.error';
 import {
   Account as DomainAccount,
   EmailAddress,
