+++ src/routes/email/exception-filters/acocunt-does-not-exist.exception-filter.ts
@@ -5,7 +5,7 @@ import {
   HttpStatus,
 } from '@nestjs/common';
 import { Response } from 'express';
-import { AccountDoesNotExistError } from '@/datasources/account/errors/account-does-not-exist.error';
+import { AccountDoesNotExistError } from '@/domain/account/errors/account-does-not-exist.error';
 
 @Catch(AccountDoesNotExistError)
 export class AccountDoesNotExistExceptionFilter implements ExceptionFilter {
